Revert "Just populate the minimum vault variables"
This reverts commit 42145dc9a149a3719054b40e805cdd9b10ed296a.

The secret CSI driver works by mounting secrets during pod creation, which does not work well with FluxCD that appears to use them in a more ad-hoc manner.
I did try making an empty pod to force creation of the secret, but unsure why it didn't work at this point in time (getting lots of deadline exceeded errors, but didn't did that deep due to the time).
Ref: https://fluxcd.io/flux/security/secrets-management/#secrets-mounted-via-csi-drivers